requirements.txt
torch==2.8.0
torchvision==0.14.1
numpy==1.26.4
Pillow
Flask
gunicorn
opencv-python-headless
